requirements.txt
streamlit
plotly
datetime
pandas 
requests
seaborn
python-dotenv